Update base image, serve file simply to be able to see result
Update base image to balenalib, as that's the future (present).

Using the 'serve' package https://www.npmjs.com/package/serve to
quickly show the single image shot, over the network or public URL.
This is done, as otherwise the project doesn't work as described in
the readme - it will loop as the camera exits, and one can never
really get into theck `/data`. With the file server it will stay
on and people can see the image itself, unlike just seeing there's
a file, when connecting over the web terminal. Thus the promise
of the readme should be better fulfilled.

Also add current workaround for the issue in #3, downgrading
the packages that are required for `raspistill`. When all our
Pi images have updated firmware, we'll be able to remove it (and
should remove it at that time).

Change-type: minor
Signed-off-by: Gergely Imreh <gergely@balena.io>